wrote fetchPotentialMatches in redux thunk
