+++ src/store/usersReducer.js
@@ -1,9 +1,15 @@
 import { firebase } from '../firebaseSpecs/config'
 
-const SET_ALL_USERS = 'SET_ALL_USERS'
+const SET_DISCOVER_USERS = 'SET_ALL_USERS'
+const SET_POTENTIAL_MATCHES = 'SET_POTENTIAL_MATCHES'
 
-export const setAllUsers = (users) => ({
-  type: SET_ALL_USERS,
+export const setDiscoverUsers = (users) => ({
+  type: SET_DISCOVER_USERS,
+  users,
+})
+
+export const setPotentialMatches = (users) => ({
+  type: SET_POTENTIAL_MATCHES,
   users,
 })
 
@@ -19,14 +25,38 @@ export const fetchUsersWithInterests = (interests) => {
           const id = doc.id
           return { id, ...data }
         })
-        dispatch(setAllUsers(users))
+        dispatch(setDiscoverUsers(users))
+      })
+  }
+}
+
+export const fetchPotentialMatches = (id) => {
+  return async (dispatch) => {
+    const userRef = firebase.firestore().collection('users')
+    await userRef
+      .where('likes', 'array-contains', id)
+      .get()
+      .then((snapshot) => {
+        if (snapshot) {
+          let users = snapshot.docs.map((doc) => {
+            const data = doc.data()
+            const id = doc.id
+            return { id, ...data }
+          })
+          dispatch(setPotentialMatches(users))
+          console.log('>>>>>> Users pulled from firestore: ', users)
+        } else {
+          console.log('users do not exist')
+        }
       })
   }
 }
 
 export default function (state = [], action) {
   switch (action.type) {
-    case SET_ALL_USERS:
+    case SET_DISCOVER_USERS:
+      return action.users
+    case SET_POTENTIAL_MATCHES:
       return action.users
     default:
       return state
